seperate game component to its own file
