+++ src/index.css
@@ -24,6 +24,7 @@ body {
   /* margin: 0;
   display: flex;
   place-items: center; */
+  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   padding: 2rem;
   min-width: 320px;
   min-height: 100vh;
